Move wms imports to editor settings
(cherry picked from commit 41517893a4f1c3d89358557fa4475026da7befce)
